% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_prettyprint.R
\name{app_prettyprint}
\alias{app_prettyprint}
\title{Shiny Interface for {XlsFormPrettyPrint} pretty output}
\usage{
app_prettyprint()
}
\value{
tart the app
}
\description{
This builds on the template and function
}
\examples{
#app_prettyprint()
}
