% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flextable.R
\name{create_flextable}
\alias{create_flextable}
\title{Generate a flextable from the pretty printed version of the xlsform}
\usage{
create_flextable(xlsformpath)
}
\arguments{
\item{xlsformpath}{}
}
\value{
a flextable
}
\description{
Generate a flextable from the pretty printed version of the xlsform
}
\examples{
create_flextable(xlsformpath = system.file("demo.xlsx", package = "XlsFormPrettyPrint") )
}
