% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview_duration.R
\name{interview_duration}
\alias{interview_duration}
\title{Prepare a summary estimation of the interview duration.}
\usage{
interview_duration(
  xlsformpath,
  label_language = NULL,
  wpm = 180,
  maxmodalities = 7,
  resptimeclose = 4,
  resptimecondopen = 7,
  resptimeopen = 10,
  avrgrepeat = 3
)
}
\arguments{
\item{xlsformpath}{path to the file with xlsform}

\item{label_language}{Language to be used in case you have more than one.
If not specified, the 'default_language' in the 'settings' worksheet is used.
If that is not specified and more than one language is in the XlsForm,
the language that comes first within column order will be used.}

\item{wpm}{word per minute - an average 180 word per minute (per default)
required to read loudly the text}

\item{maxmodalities}{if more than 7 potential answers for a select question
(per default)- then we assume that those modalities will not be read by the
enumerator - but rather selected based on an open answer - and not be
accounted for the modalities duration estimation}

\item{resptimeclose}{an average 4 seconds (per default) for respondent to
reply for closed questions}

\item{resptimecondopen}{an average of  7 seconds (per default) to reply
to conditional text question (accounting for question type of
"other, please specify").}

\item{resptimeopen}{an average of  10 seconds (per default) to reply to
open text question.}

\item{avrgrepeat}{In case of repeat questions, an average 3 repeat
(per default) is accounted for.}
}
\value{
a summary review in a visual format as a gpplot2 chart
}
\description{
When designing a questionnaire, it is key to keep the interview
duration under control. Typically less than 20 minutes for a phone interview
and less than 40 minutes for a face to face interview.

The function is designed to provide a rough estimate (aka a guesstimate...)
of the interview duration in order to assess this element of questionnaire
design quality. If the questionnaire takes too long, then one needs to trim
it or split it into multiple ones...

The estimations provided in that function are based on a series of assumptions
and accounts for the following elements:
\itemize{
\item the time needed to read the label of the questions - or the note -
Questions hint are not accounted for as they tips for enumerator and not
to be read to respondent
\item the time needed to read the question modalities - as there's nothing in
xlsform that tells whether the enumerator is expected to read loudly the
potential answers in case of closed question, the function has a parameter
to
\item the time need to answer the question - here we are accounting for 3
distinct situations, each of them with a different potential duration:
}
\itemize{
\item close questions - open questions and open questions set up under skip logic
so typically - the "if other, specify" - each comes with distinct response time
}
\itemize{
\item the time needed when we have a "repeat" block - typically questions repeated
for a family roster -
}
}
\examples{
## Generate the summary chart
interview_duration( 
  xlsformpath = system.file("demo.xlsx", package = "XlsFormPrettyPrint"), 
  label_language = NULL,
  # wpm  word per minute - an average 180 word per minute (per default) required to read loudly the text
  wpm  = 180, 
# maxmodalities if more than 7 potential answers for a select question (per default)- then we assume that those modalities will not be read by the enumerator - but rather selected based on an open answer - and not be accounted for the modalities duration estimation
  maxmodalities = 7 , 
# resptimeclose  an average 4 seconds (per default) for respondent to reply for closed questions
  resptimeclose  = 4,
# resptimecondopen an average of  7 seconds (per default) to reply to conditional text question (accounting for question type of "other, please specify"). 
  resptimecondopen = 7,
# resptimeopen an average of  10 seconds (per default) to reply to open text question. 
  resptimeopen = 10,
# avrgrepeat In case of repeat questions, an average 3 repeat (per default) is accounted for. 
  avrgrepeat = 3 
  )


}
