% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_prettyprint.R
\name{render_prettyprint}
\alias{render_prettyprint}
\title{Render Word version of an xlsform}
\usage{
render_prettyprint(
  dir,
  xlsformfile,
  label_language = NULL,
  wpm = 180,
  maxmodalities = 7,
  resptimeclose = 4,
  resptimecondopen = 7,
  resptimeopen = 10,
  avrgrepeat = 3
)
}
\arguments{
\item{dir}{directory where the xlsform is located}

\item{xlsformfile}{name of the file for the xlsform}

\item{label_language}{language to sue to generate the pretty print}

\item{wpm}{word per minute - an average 180 word per minute (per default)
required to read loudly the text}

\item{maxmodalities}{if more than 7 potential answers for a select question
(per default)- then we assume that those modalities will not be read by the
enumerator - but rather selected based on an open answer - and not be
accounted for the modalities duration estimation}

\item{resptimeclose}{an average 4 seconds (per default) for respondent
to reply for closed questions}

\item{resptimecondopen}{an average of  7 seconds (per default) to reply
to conditional text question (accounting for question type of
"other, please specify").}

\item{resptimeopen}{an average of  10 seconds (per default) to reply
to open text question.}

\item{avrgrepeat}{In case of repeat questions, an average 3 repeat
(per default) is accounted for.}
}
\value{
a word document
}
\description{
Us the built in template and functions from the package to generate a word
version of the xlsform
}
\examples{
# render_prettyprint( dir =  system.file("", package = "XlsFormPrettyPrint"),
#   xlsformfile = "demo.xlsx",
#   label_language = NULL )
}
