% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_form.R
\name{tabulate_form}
\alias{tabulate_form}
\title{Combine the xlsform in single table for pretty printing}
\usage{
tabulate_form(xlsformpath)
}
\arguments{
\item{xlsformpath}{}
}
\value{
a single data frame
}
\description{
Combine the xlsform in single table for pretty printing
}
\examples{
# tabulate_form()

prettyform <- tabulate_form( xlsformpath = system.file("demo.xlsx", package = "XlsFormPrettyPrint") )

knitr::kable(utils::head(prettyform, 10))

}
