% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_form.R
\name{tabulate_form}
\alias{tabulate_form}
\title{Combine the xlsform in single table for pretty printing}
\usage{
tabulate_form(xlsformpath, label_language = NULL)
}
\arguments{
\item{xlsformpath}{path to the file with xlsform}

\item{label_language}{Language to be used in case you have more than one. If not specified, the 'default_language' in the 'settings' worksheet is used. If that is not specified and more than one language is in the XlsForm, the language that comes first within column order will be used.}
}
\value{
a single data frame
}
\description{
Combine the xlsform in single table for pretty printing
}
\examples{
# tabulate_form()

prettyform <- tabulate_form( xlsformpath = system.file("demo.xlsx", package = "XlsFormPrettyPrint") )

knitr::kable(utils::head(prettyform, 10))

}
